+++ src/index.ts
@@ -1,6 +1,6 @@
 import FocusContext from './focus-context';
 export { default as FocusNode } from './focus-node';
-export { default as useFocusNode } from './hooks/use-focus-node';
+export { default as useFocusNodeById } from './hooks/use-focus-node-by-id';
 export { default as useLeafFocusedNode } from './hooks/use-leaf-focused-node';
 export { default as useActiveNode } from './hooks/use-active-node';
 export { default as useFocusHierarchy } from './hooks/use-focus-hierarchy';
